fix: consolidate regenerate scripts and fix azure spec detection
Merge regenerate-common.ts into regenerate.ts since it had only one
consumer. Also fix spec.includes('azure') which falsely matched when
the working directory contained 'azure' (e.g. azure-sdk-for-python),
causing incorrect package names for unbranded specs.
