+++ .chronus/changes/python-fix-azure-spec-detection-2026-3-29-12-59-39.md
@@ -1,0 +1,7 @@
+---
+changeKind: internal
+packages:
+  - "@typespec/http-client-python"
+---
+
+Consolidate regenerate-common.ts into regenerate.ts and fix azure spec detection using azure-http-specs path check
