+++ packages/http-client-python/eng/scripts/ci/regenerate.ts
@@ -8,21 +8,12 @@
 
 import { compile, NodeHost } from "@typespec/compiler";
 import { execSync } from "child_process";
-import { existsSync, readdirSync, rmSync } from "fs";
+import { existsSync, promises, readdirSync, rmSync } from "fs";
 import { platform } from "os";
 import { dirname, join, relative, resolve } from "path";
 import pc from "picocolors";
 import { fileURLToPath } from "url";
 import { parseArgs } from "util";
-import {
-  BASE_AZURE_EMITTER_OPTIONS,
-  BASE_EMITTER_OPTIONS,
-  getSubdirectories,
-  preprocess,
-  SpecialFlags,
-  toPosix,
-  type RegenerateFlags,
-} from "./regenerate-common.js";
 
 // Parse arguments
 const argv = parseArgs({
@@ -83,6 +74,405 @@ ${pc.bold("Examples:")}
   process.exit(0);
 }
 
+// ---- Shared constants ----
+
+const SKIP_SPECS: string[] = ["type/file", "service/multiple-services"];
+
+const SpecialFlags: Record<string, Record<string, any>> = {
+  azure: {
+    "generate-test": true,
+    "generate-sample": true,
+  },
+};
+
+function toPosix(dir: string): string {
+  return dir.replace(/\\/g, "/");
+}
+
+interface RegenerateFlags {
+  flavor: string;
+  debug: boolean;
+  name?: string;
+  pyodide?: boolean;
+}
+
+// ---- Base emitter options ----
+
+const AZURE_EMITTER_OPTIONS: Record<
+  string,
+  Record<string, string> | Record<string, string>[]
+> = {
+  "azure/client-generator-core/access": {
+    namespace: "specs.azure.clientgenerator.core.access",
+  },
+  "azure/client-generator-core/alternate-type": {
+    namespace: "specs.azure.clientgenerator.core.alternatetype",
+  },
+  "azure/client-generator-core/api-version": {
+    namespace: "specs.azure.clientgenerator.core.apiversion",
+  },
+  "azure/client-generator-core/client-initialization/default": {
+    namespace: "specs.azure.clientgenerator.core.clientinitialization.default",
+  },
+  "azure/client-generator-core/client-initialization/individually": {
+    namespace: "specs.azure.clientgenerator.core.clientinitialization.individually",
+  },
+  "azure/client-generator-core/client-initialization/individuallyParent": {
+    namespace: "specs.azure.clientgenerator.core.clientinitialization.individuallyparent",
+  },
+  "azure/client-generator-core/client-location": {
+    namespace: "specs.azure.clientgenerator.core.clientlocation",
+  },
+  "azure/client-generator-core/deserialize-empty-string-as-null": {
+    namespace: "specs.azure.clientgenerator.core.emptystring",
+  },
+  "azure/client-generator-core/flatten-property": {
+    namespace: "specs.azure.clientgenerator.core.flattenproperty",
+  },
+  "azure/client-generator-core/usage": {
+    namespace: "specs.azure.clientgenerator.core.usage",
+  },
+  "azure/client-generator-core/client-doc": {
+    namespace: "specs.azure.clientgenerator.core.clientdoc",
+  },
+  "azure/client-generator-core/override": {
+    namespace: "specs.azure.clientgenerator.core.override",
+  },
+  "azure/client-generator-core/hierarchy-building": {
+    namespace: "specs.azure.clientgenerator.core.hierarchybuilding",
+  },
+  "azure/core/basic": {
+    namespace: "specs.azure.core.basic",
+  },
+  "azure/core/lro/rpc": {
+    namespace: "specs.azure.core.lro.rpc",
+  },
+  "azure/core/lro/standard": {
+    namespace: "specs.azure.core.lro.standard",
+  },
+  "azure/core/model": {
+    namespace: "specs.azure.core.model",
+  },
+  "azure/core/page": {
+    namespace: "specs.azure.core.page",
+  },
+  "azure/core/scalar": {
+    namespace: "specs.azure.core.scalar",
+  },
+  "azure/core/traits": {
+    namespace: "specs.azure.core.traits",
+  },
+  "azure/encode/duration": {
+    namespace: "specs.azure.encode.duration",
+  },
+  "azure/example/basic": {
+    namespace: "specs.azure.example.basic",
+  },
+  "azure/payload/pageable": {
+    namespace: "specs.azure.payload.pageable",
+  },
+  "azure/versioning/previewVersion": {
+    namespace: "specs.azure.versioning.previewversion",
+  },
+  "client/structure/default": {
+    namespace: "client.structure.service",
+  },
+  "client/structure/multi-client": {
+    "package-name": "client-structure-multiclient",
+    namespace: "client.structure.multiclient",
+  },
+  "client/structure/renamed-operation": {
+    "package-name": "client-structure-renamedoperation",
+    namespace: "client.structure.renamedoperation",
+  },
+  "client/structure/two-operation-group": {
+    "package-name": "client-structure-twooperationgroup",
+    namespace: "client.structure.twooperationgroup",
+  },
+  "client/naming": {
+    namespace: "client.naming.main",
+  },
+  "client/overload": {
+    namespace: "client.overload",
+  },
+  "encode/duration": {
+    namespace: "encode.duration",
+  },
+  "encode/numeric": {
+    namespace: "encode.numeric",
+  },
+  "parameters/basic": {
+    namespace: "parameters.basic",
+  },
+  "parameters/spread": {
+    namespace: "parameters.spread",
+  },
+  "payload/content-negotiation": {
+    namespace: "payload.contentnegotiation",
+  },
+  "payload/multipart": {
+    namespace: "payload.multipart",
+  },
+  "serialization/encoded-name/json": {
+    namespace: "serialization.encodedname.json",
+  },
+  "special-words": {
+    namespace: "specialwords",
+  },
+  "service/multi-service": {
+    namespace: "service.multiservice",
+  },
+  "client/structure/client-operation-group": {
+    "package-name": "client-structure-clientoperationgroup",
+    namespace: "client.structure.clientoperationgroup",
+  },
+};
+
+const EMITTER_OPTIONS: Record<
+  string,
+  Record<string, string> | Record<string, string>[]
+> = {
+  "resiliency/srv-driven/old.tsp": {
+    "package-name": "resiliency-srv-driven1",
+    namespace: "resiliency.srv.driven1",
+    "package-mode": "azure-dataplane",
+    "package-pprint-name": "ResiliencySrvDriven1",
+  },
+  "resiliency/srv-driven": {
+    "package-name": "resiliency-srv-driven2",
+    namespace: "resiliency.srv.driven2",
+    "package-mode": "azure-dataplane",
+    "package-pprint-name": "ResiliencySrvDriven2",
+  },
+  "authentication/api-key": {
+    "clear-output-folder": "true",
+  },
+  "authentication/http/custom": {
+    "package-name": "authentication-http-custom",
+    namespace: "authentication.http.custom",
+    "package-pprint-name": "Authentication Http Custom",
+  },
+  "authentication/union": [
+    {
+      "package-name": "authentication-union",
+      namespace: "authentication.union",
+    },
+    {
+      "package-name": "setuppy-authentication-union",
+      namespace: "setuppy.authentication.union",
+      "keep-setup-py": "true",
+    },
+  ],
+  "type/array": {
+    "package-name": "typetest-array",
+    namespace: "typetest.array",
+  },
+  "type/dictionary": {
+    "package-name": "typetest-dictionary",
+    namespace: "typetest.dictionary",
+  },
+  "type/enum/extensible": {
+    "package-name": "typetest-enum-extensible",
+    namespace: "typetest.enum.extensible",
+  },
+  "type/enum/fixed": {
+    "package-name": "typetest-enum-fixed",
+    namespace: "typetest.enum.fixed",
+  },
+  "type/model/empty": {
+    "package-name": "typetest-model-empty",
+    namespace: "typetest.model.empty",
+  },
+  "type/model/inheritance/enum-discriminator": {
+    "package-name": "typetest-model-enumdiscriminator",
+    namespace: "typetest.model.enumdiscriminator",
+  },
+  "type/model/inheritance/nested-discriminator": {
+    "package-name": "typetest-model-nesteddiscriminator",
+    namespace: "typetest.model.nesteddiscriminator",
+  },
+  "type/model/inheritance/not-discriminated": {
+    "package-name": "typetest-model-notdiscriminated",
+    namespace: "typetest.model.notdiscriminated",
+  },
+  "type/model/inheritance/single-discriminator": {
+    "package-name": "typetest-model-singlediscriminator",
+    namespace: "typetest.model.singlediscriminator",
+  },
+  "type/model/inheritance/recursive": [
+    {
+      "package-name": "typetest-model-recursive",
+      namespace: "typetest.model.recursive",
+    },
+    {
+      "package-name": "generation-subdir",
+      namespace: "generation.subdir",
+      "generation-subdir": "_generated",
+      "generate-test": "false",
+      "clear-output-folder": "true",
+    },
+  ],
+  "type/model/usage": {
+    "package-name": "typetest-model-usage",
+    namespace: "typetest.model.usage",
+  },
+  "type/model/visibility": [
+    {
+      "package-name": "typetest-model-visibility",
+      namespace: "typetest.model.visibility",
+    },
+    {
+      "package-name": "headasbooleantrue",
+      namespace: "headasbooleantrue",
+      "head-as-boolean": "true",
+    },
+    {
+      "package-name": "headasbooleanfalse",
+      namespace: "headasbooleanfalse",
+      "head-as-boolean": "false",
+    },
+  ],
+  "type/property/nullable": {
+    "package-name": "typetest-property-nullable",
+    namespace: "typetest.property.nullable",
+  },
+  "type/property/optionality": {
+    "package-name": "typetest-property-optional",
+    namespace: "typetest.property.optional",
+  },
+  "type/property/additional-properties": {
+    "package-name": "typetest-property-additionalproperties",
+    namespace: "typetest.property.additionalproperties",
+  },
+  "type/scalar": {
+    "package-name": "typetest-scalar",
+    namespace: "typetest.scalar",
+  },
+  "type/property/value-types": {
+    "package-name": "typetest-property-valuetypes",
+    namespace: "typetest.property.valuetypes",
+  },
+  "type/union": {
+    "package-name": "typetest-union",
+    namespace: "typetest.union",
+  },
+  "type/union/discriminated": {
+    "package-name": "typetest-discriminatedunion",
+    namespace: "typetest.discriminatedunion",
+  },
+  "type/file": {
+    "package-name": "typetest-file",
+    namespace: "typetest.file",
+  },
+  documentation: {
+    "package-name": "specs-documentation",
+    namespace: "specs.documentation",
+  },
+  "versioning/added": [
+    {
+      "package-name": "versioning-added",
+      namespace: "versioning.added",
+    },
+    {
+      "package-name": "generation-subdir2",
+      namespace: "generation.subdir2",
+      "generate-test": "false",
+      "generation-subdir": "_generated",
+    },
+  ],
+};
+
+// ---- Shared utility functions ----
+
+async function getSubdirectories(
+  baseDir: string,
+  flags: RegenerateFlags,
+): Promise<string[]> {
+  const subdirectories: string[] = [];
+
+  async function searchDir(currentDir: string) {
+    const items = await promises.readdir(currentDir, { withFileTypes: true });
+
+    const promisesArray = items.map(async (item) => {
+      const subDirPath = join(currentDir, item.name);
+      if (item.isDirectory()) {
+        const mainTspPath = join(subDirPath, "main.tsp");
+        const clientTspPath = join(subDirPath, "client.tsp");
+
+        const mainTspRelativePath = toPosix(relative(baseDir, mainTspPath));
+
+        if (SKIP_SPECS.some((skipSpec) => mainTspRelativePath.includes(skipSpec))) return;
+
+        const hasMainTsp = await promises
+          .access(mainTspPath)
+          .then(() => true)
+          .catch(() => false);
+        const hasClientTsp = await promises
+          .access(clientTspPath)
+          .then(() => true)
+          .catch(() => false);
+
+        if (mainTspRelativePath.toLowerCase().includes(flags.name || "")) {
+          if (mainTspRelativePath.includes("resiliency/srv-driven")) {
+            subdirectories.push(resolve(subDirPath, "old.tsp"));
+          }
+          if (hasClientTsp) {
+            subdirectories.push(resolve(subDirPath, "client.tsp"));
+          } else if (hasMainTsp) {
+            subdirectories.push(resolve(subDirPath, "main.tsp"));
+          }
+        }
+
+        await searchDir(subDirPath);
+      }
+    });
+
+    await Promise.all(promisesArray);
+  }
+
+  await searchDir(baseDir);
+  return subdirectories;
+}
+
+async function preprocess(flavor: string, generatedFolder: string): Promise<void> {
+  if (flavor === "azure") {
+    const testsGeneratedDir = resolve(generatedFolder, "../tests/generated/azure");
+
+    const DELETE_CONTENT = "# This file is to be deleted after regeneration";
+    const DELETE_FILE = "to_be_deleted.py";
+    const entries: { folder: string[]; file: string; content: string }[] = [
+      {
+        folder: ["authentication-api-key", "authentication", "apikey", "_operations"],
+        file: DELETE_FILE,
+        content: DELETE_CONTENT,
+      },
+      {
+        folder: ["generation-subdir", "generation", "subdir", "_generated"],
+        file: DELETE_FILE,
+        content: DELETE_CONTENT,
+      },
+      {
+        folder: ["generation-subdir", "generated_tests"],
+        file: DELETE_FILE,
+        content: DELETE_CONTENT,
+      },
+      {
+        folder: ["generation-subdir", "generation", "subdir"],
+        file: "to_be_kept.py",
+        content: "# This file is to be kept after regeneration",
+      },
+    ];
+
+    await Promise.all(
+      entries.map(async ({ folder, file, content }) => {
+        const targetFolder = join(testsGeneratedDir, ...folder);
+        await promises.mkdir(targetFolder, { recursive: true });
+        await promises.writeFile(join(targetFolder, file), content);
+      }),
+    );
+  }
+}
+
 // Get paths
 const SCRIPT_DIR = dirname(fileURLToPath(import.meta.url));
 const PLUGIN_DIR = argv.values.pluginDir
@@ -94,23 +484,6 @@ const GENERATED_FOLDER = argv.values.generatedFolder
   ? resolve(argv.values.generatedFolder)
   : resolve(PLUGIN_DIR, "generator");
 const EMITTER_NAME = argv.values.emitterName || "@typespec/http-client-python";
-
-// Emitter options
-const AZURE_EMITTER_OPTIONS: Record<string, Record<string, string> | Record<string, string>[]> = {
-  ...BASE_AZURE_EMITTER_OPTIONS,
-  "client/structure/client-operation-group": {
-    "package-name": "client-structure-clientoperationgroup",
-    namespace: "client.structure.clientoperationgroup",
-  },
-};
-
-const EMITTER_OPTIONS: Record<string, Record<string, string> | Record<string, string>[]> = {
-  ...BASE_EMITTER_OPTIONS,
-  "type/array": {
-    "package-name": "typetest-array",
-    namespace: "typetest.array",
-  },
-};
 
 interface CompileTask {
   spec: string;
